Move base `RustSymbolProvider` types out of `SymbolVisitor`
