+++ codegen-core/src/main/kotlin/software/amazon/smithy/rust/codegen/core/smithy/SymbolVisitor.kt
@@ -66,27 +66,6 @@ val SimpleShapes: Map<KClass<out Shape>, RustType> = mapOf(
 )
 
 /**
- * Provider for RustModules so that the symbol provider knows where to organize things.
- */
-interface ModuleProvider {
-    /** Returns the module for a shape */
-    fun moduleForShape(shape: Shape): RustModule.LeafModule
-
-    /** Returns the module for an operation error */
-    fun moduleForOperationError(operation: OperationShape): RustModule.LeafModule
-
-    /** Returns the module for an event stream error */
-    fun moduleForEventStreamError(eventStream: UnionShape): RustModule.LeafModule
-}
-
-data class SymbolVisitorConfig(
-    val runtimeConfig: RuntimeConfig,
-    val renameExceptions: Boolean,
-    val nullabilityCheckMode: CheckMode,
-    val moduleProvider: ModuleProvider,
-)
-
-/**
  * Make the Rust type of a symbol optional (hold `Option<T>`)
  *
  * This is idempotent and will have no change if the type is already optional.
@@ -173,26 +152,6 @@ fun Symbol.Builder.locatedIn(rustModule: RustModule.LeafModule): Symbol.Builder 
  * field will be set.
  */
 data class MaybeRenamed(val name: String, val renamedFrom: String?)
-
-/**
- * SymbolProvider interface that carries both the inner configuration and a function to produce an enum variant name.
- */
-interface RustSymbolProvider : SymbolProvider, ModuleProvider {
-    fun config(): SymbolVisitorConfig
-    fun toEnumVariantName(definition: EnumDefinition): MaybeRenamed?
-
-    override fun moduleForShape(shape: Shape): RustModule.LeafModule = config().moduleProvider.moduleForShape(shape)
-    override fun moduleForOperationError(operation: OperationShape): RustModule.LeafModule =
-        config().moduleProvider.moduleForOperationError(operation)
-    override fun moduleForEventStreamError(eventStream: UnionShape): RustModule.LeafModule =
-        config().moduleProvider.moduleForEventStreamError(eventStream)
-
-    /** Returns the symbol for an operation error */
-    fun symbolForOperationError(operation: OperationShape): Symbol
-
-    /** Returns the symbol for an event stream error */
-    fun symbolForEventStreamError(eventStream: UnionShape): Symbol
-}
 
 /**
  * Make the return [value] optional if the [member] symbol is as well optional.
